client_requirements.txt
python-socketio
aiohttp
PyGObject
asyncio-glib
